Cleanup and safely update dependencies
Clearly it is not the best practice to leave thousands
of security vulnerabilities in the wild, especially
when there are two "critical" ones with "Remote Arbitrary
Code Execution" capability.

This change updates dependencies safely and fixes
all vulnerabilities above "low" severity.

test.js and its dependencies are removed as there is
no test and it seems that there will never be one. It
can be added back later if tests *are* implemented.

All functions in "npm run" are tested. At the first glance,
there is no issue.

Signed-off-by: Jesse Chan <jc@linux.com>