client: import flood-ui-kit
flood-ui-kit is highly tied to flood and no other project uses
it for now. It doesn't worth the effort to maintain it in a
separate package.

This allows us to more easily tweak the UI and sync the changes.

Repo: jfurrow/flood-ui-kit
Revision: 42d472085d127e58602595921c6279106c41fc7d
Signed-off-by: Jesse Chan <jc@linux.com>