Tooltip: simplify Portal creation
It is unneccessary and unreliable to create an extra div to hold Portal. The
presumed benefits[1] of an extra div are never proven and developers never
find an *ACTUAL* case where putting Portal to the body poses a problem[2].

In our case, creation of this extra div (or the logics to do so) is the
*ACTUAL* problem. Somehow the "ref" inside this div doesn't work reliably.
As a result, getIdealLocation and handleMouseEnter fail.

[1]: https://stackoverflow.com/questions/49504546/is-it-safe-to-use-reactdom-createportal-with-document-body
[2]: https://github.com/mui-org/material-ui/issues/21626

Bug: #15