feature: allow to replace main tracker of torrents
rTorrent does not have function to remove trackers from torrents
so we only disable existing tracker and insert the new one.

This leads to duplicate listings of trackers in "Trackers" and
"Filter by trackers". We can fix this later by only using enabled
trackers for "Filter by" and "Torrent Details".

Currently this feature only deals with the main tracker (index: 0).
Combined with "Filter by trackers", it is ideal for a bulk change
of tracker URL of private torrents due to a key leak or rotation.

In the future, we can probably make the feature more sophisticated.