ClientStatusStore: assume backend's connection with rTorrent is OK
This flag indicates connectivity between backend (Flood server) and
rTorrent's SCGI interface. Before this change, we assume "not OK".

However, when frondend is authenticated and is in the process of
fetching settings from the backend. Connection interrupted page
will always show momentarily given the conditionals of isAuthenticated
&& !isConnected even if the connection is perfectly OK.

This change makes frontend assume the backend's connection with rTorrent
is OK by default.